@@ -4,7 +4,6 @@
   "description": "SQLite database distribution of airport data for the airport-db ecosystem",
   "type": "module",
   "sideEffects": false,
-  "main": "./index.js",
   "homepage": "https://github.com/tabletopandroid/airport-db-data-sqlite",
   "repository": {
     "type": "git",
@@ -14,23 +13,32 @@
     "url": "https://github.com/tabletopandroid/airport-db-data-sqlite/issues"
   },
   "engines": {
-    "node": ">=14.0.0"
+    "node": ">=18.0.0"
   },
   "files": [
-    "data",
-    "index.js",
+    "dist",
     "README.md",
     "CONTRIBUTING.md",
     "LICENSE"
   ],
   "exports": {
     ".": {
-      "default": "./index.js"
+      "default": "./dist/index.js"
     }
   },
   "scripts": {
+    "build": "npm run copy-entry",
+    "build:full": "npm run generate-db && npm run import-data && npm run copy-entry",
+    "changeset": "changeset",
+    "version-packages": "changeset version",
+    "publish-packages": "changeset publish",
+    "copy-entry": "node scripts/build-entry.js",
     "generate-db": "node scripts/generate-db.js",
-    "import-data": "node scripts/import-csv.cjs"
+    "import-data": "node scripts/import-csv.cjs",
+    "release:github": "node scripts/github-release.js",
+    "verify:dist-db": "node -e \"import fs from 'node:fs'; if (!fs.existsSync('dist/airports.sqlite')) { console.error('Missing dist/airports.sqlite. Run npm run build:full first.'); process.exit(1); }\"",
+    "prepublishOnly": "npm run verify:dist-db && npm run build",
+    "postpublish": "npm run release:github"
   },
   "keywords": [
     "airport",
@@ -46,6 +54,7 @@
   "author": "tabletop ANDROID",
   "license": "ODbL-1.0",
   "devDependencies": {
+    "@changesets/cli": "^2.29.8",
     "sql.js": "^1.14.0"
   }
 }
